adapt day2

+++ day2/src/main.rs
@@ -11,7 +11,7 @@ fn main() {
 
     println!("Intcode Return: {:?}", intcode);
 
-    let valid_values: Vec<usize> = find_inputs_for(&original_code, 19_690_720)
+    let valid_values: Vec<i32> = find_inputs_for(&original_code, 19_690_720)
         .iter()
         .map(|(noun, verb)| 100 * noun + verb)
         .collect();
@@ -19,7 +19,7 @@ fn main() {
     println!("valid inputs are: {:?}", valid_values)
 }
 
-fn find_inputs_for(memory: &Memory, wanted_output: usize) -> Vec<(usize, usize)> {
+fn find_inputs_for(memory: &Memory, wanted_output: i32) -> Vec<(i32, i32)> {
     let mut valid_values = vec![];
     for (noun, verb) in iproduct!(0..99, 0..99) {
         let code = get_custom_inputs(&memory, noun, verb);
@@ -43,7 +43,7 @@ fn find_inputs_for(memory: &Memory, wanted_output: usize) -> Vec<(usize, usize)>
     valid_values
 }
 
-fn get_custom_inputs(memory: &Memory, noun: usize, verb: usize) -> Memory {
+fn get_custom_inputs(memory: &Memory, noun: i32, verb: i32) -> Memory {
     let mut new_memory = memory.clone();
 
     new_memory[1] = noun;
